Added order, pending test

+++ src/app/admin/positions/position.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { ModalService } from 'src/app/services/modal.service';
 import { Position } from 'src/app/models/position.model';
-import { FormGroup, FormBuilder, Validators, FormArray, FormControl } from '@angular/forms';
+import { FormGroup, FormBuilder, Validators, FormArray, FormControl, ValidatorFn, AbstractControl } from '@angular/forms';
 import { SkillType } from 'src/app/models/skill.model';
+import { CommonsService } from 'src/app/services/commons.service';
 
 @Component({
   selector: 'app-position',
@@ -18,6 +19,7 @@ export class PositionComponent implements OnInit {
 
   constructor(
     private modalService: ModalService,
+    private commonsService: CommonsService,
     private fb: FormBuilder
   ) {
     this.positionform = this.fb.group({
@@ -31,7 +33,9 @@ export class PositionComponent implements OnInit {
       skills: [''],
       normal: new FormArray([]),
       doubles: new FormArray([])
-    });
+      },
+      { validator: this.validateTypes }
+    );
 
     document.getElementById('modal-div').style.maxWidth = '960px';
   }
@@ -70,14 +74,69 @@ export class PositionComponent implements OnInit {
   public close() {
     this.modalService.setOutput(false);
     document.getElementById('modal-div').style.maxWidth = '';
-    // document.getElementsByClassName('modal')[0].classList.remove('modal-lg');
     this.modalService.destroy();
   }
 
-  public confirm() {
-    this.modalService.setOutput(true);
-    document.getElementById('modal-div').style.maxWidth = '';
-    // document.getElementsByClassName('modal')[0].classList.remove('modal-lg');
-    this.modalService.destroy();
+  public onSubmit() {
+    if (this.positionform.valid) {
+      this.modalService.setOutput(true);
+      document.getElementById('modal-div').style.maxWidth = '';
+      this.modalService.destroy();
+    } else {
+      this.commonsService.handleError('Hay campos sin rellenar');
+    }
+  }
+
+  get name() {
+    return this.positionform.get('name');
+  }
+  get limit() {
+    return this.positionform.get('limit');
+  }
+  get price() {
+    return this.positionform.get('price');
+  }
+  get ma() {
+    return this.positionform.get('ma');
+  }
+  get st() {
+    return this.positionform.get('st');
+  }
+  get ag() {
+    return this.positionform.get('ag');
+  }
+  get av() {
+    return this.positionform.get('av');
+  }
+  get normal() {
+    return this.positionform.get('normal').value;
+  }
+  get doubles() {
+    return this.positionform.get('doubles').value;
+  }
+
+  validateTypes(group: FormGroup): any {
+    if (group) {
+      // Validar que la suma de ambos arrays es 6 entre los 2 y que no se repiten
+      const totalNormal = this.normal ? this.normal
+        // get a list of checkbox values (boolean)
+        .map(control => control.value)
+        // total up the number of checked checkboxes
+        .reduce((prev, next) => next ? prev + next : prev, 0) : 0;
+
+      const totalDoubles = this.doubles ? this.doubles
+        // get a list of checkbox values (boolean)
+        .map(control => control.value)
+        // total up the number of checked checkboxes
+        .reduce((prev, next) => next ? prev + next : prev, 0) : 0;
+
+      if (totalNormal + totalDoubles > 6 || totalNormal + totalDoubles < 5) {
+        return { wrongNumber: true };
+      }
+
+      return null;
+    }
+
+    return null;
   }
 }
